% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapTests-package.R
\docType{package}
\name{BootstrapTests-package}
\alias{BootstrapTests}
\alias{BootstrapTests-package}
\title{BootstrapTests: Use Several Methods to Perform Bootstrap-Based Hypothesis Testing Procedures}
\description{
Perform bootstrap-based hypothesis testing procedures on three statistical problems. In particular, it covers independence testing, testing the slope in a linear regression setting, and goodness-of-fit testing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AlexisDerumigny/BootstrapTests}
  \item Report bugs at \url{https://github.com/AlexisDerumigny/BootstrapTests/issues}
}

}
\author{
\strong{Maintainer}: Wieger Schipper \email{W.R.Schipper@tudelft.nl} (\href{https://orcid.org/0009-0004-5661-4949}{ORCID})

Authors:
\itemize{
  \item Alexis Derumigny \email{A.F.F.Derumigny@tudelft.nl} (\href{https://orcid.org/0000-0002-6163-8097}{ORCID})
  \item Miltiadis Galanis \email{mgalan@di.uoa.gr}
  \item Aad van der Vaart \email{A.W.vanderVaart@tudelft.nl} (\href{https://orcid.org/0000-0002-8074-2375}{ORCID})
}

}
\keyword{internal}
