% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_GoF_test.R
\name{perform_GoF_test}
\alias{perform_GoF_test}
\title{Perform a GoF test}
\usage{
perform_GoF_test(
  X_data,
  parametric_fam = "normal",
  nBootstrap = 100,
  type_boot_user = "null",
  type_stat_user = "eq",
  param_bs_user = "MD"
)
}
\arguments{
\item{X_data, }{numerical input vector. Perform a GoF test whether or not this
sample comes from `parametric_fam`, a specified parametric distribution.}

\item{parametric_fam}{name of the parametric family. For the moment, only
\code{"normal"} is supported.}

\item{nBootstrap}{numeric value of the number of bootstrap resamples. Defaults
to 100.}

\item{type_boot_user}{defaults to the "null" bootstrap resampling scheme to be used.
\code{type_boot_user} can be either "null" for the null/parametric bootstrap,
or \code{"NP"} for the non-parametric bootstrap.}

\item{type_stat_user}{defaults to \code{"eq"} for the type of test statistic
to be used. This can be either \code{"eq"} for the equivalent test statistic,
or \code{"cent"} for the centered test statistic.}

\item{param_bs_user}{defaults to \code{"MD"} for the bootstrap parameter
estimator to be used. \code{param_bs_user} can be either \code{"MD"} for the
Minimum Distance estimator, \code{"MD-cent"} for the centered Minimum Distance
estimator, or \code{"canonical"} for the canonical estimator.}
}
\value{
A class object with components \itemize{
   \item \code{pvals_df} df of p-values and bootstrapped test statistics:

   These are the p-values for the combinations of bootstrap resampling schemes,
   test statistics (centered and equivalent), and different parameter estimators.

   It also contains the vectors of bootstrap test statistics
   for each of the combinations.

   \item \code{true_stat} a named vector of size 2 containing the true test
   statistics. The first entry is the Kolmogorov-Smirnov test statistic for
   the Minimum Distance estimator, and the second entry is the Kolmogorov-Smirnov
   test statistic for the canonical parameter estimator.

   \item \code{nBootstrap} number of bootstrap repetitions.

   \item \code{highlighted_pval} a dataframe with the default GoF test
   results.

   \item \code{nameMethod} string for the name of the method used.

}
}
\description{
This function performs a bootstrap goodness-of-fit test for a specific univariate
parametric family. It implements a null bootstrap and a non-parametric bootstrap.
The test statistic is the Kolmogorov-Smirnov test statistic.
For the null/parametric bootstrap, the minimum distance estimator is used to
estimate the parameters, or a canonical estimator (the sample mean and variance).
For now, only a test of normality is implemented.
}
\examples{
n = 100

# Under H1
X_data = rgamma(n,2,3)
result = perform_GoF_test(X_data, nBootstrap = 30)
print(result)
plot(result)

# Under H0
X_data = rnorm(n)
result = perform_GoF_test(X_data, nBootstrap = 30)
print(result)
plot(result)

}
\seealso{
\code{\link{perform_regression_test},\link{perform_independence_test}}
}
