% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_independence_test.R
\name{perform_independence_test}
\alias{perform_independence_test}
\title{Perform a hypothesis test of independence}
\usage{
perform_independence_test(
  X1,
  X2,
  my_grid = NULL,
  nBootstrap = 100,
  bootstrapOptions = NULL
)
}
\arguments{
\item{X1, X2}{numerical vectors of the same size. The independence test tests
whether \code{X1} is independent from \code{X2}.}

\item{my_grid}{the grid on which the empirical CDFs are estimated. It defaults
to \code{NULL}, so it is chosen automatically. However, the user can input
a grid of their choice.}

\item{nBootstrap}{number of bootstrap repetitions.}

\item{bootstrapOptions}{This can be one of \itemize{
  \item \code{NULL}

  \item a list with at most 3 elements names \itemize{
        \item \code{type_boot} defaults to the \code{"indep"} bootstrap
        resampling scheme to be used. \code{type_boot} can be either
        \code{"indep"} for the independence/null bootstrap, or \code{"NP"}
        for the non-parametric bootstrap.

        \item \code{type_stat} defaults to \code{"eq"} for the type of test
        statistic to be used. This can be either \code{"eq"} for the
        equivalent test statistic, or \code{"cent"} for the centered
        test statistic.

        \item \code{norm_type} defaults to \code{"KS"} for the type of norm
        to bes used for the test statistic. \code{norm_type} can be either
        \code{"KS"} for the Kolmogorov-Smirnov type test statistic or
        \code{"L2"} for the L2-type test statistic.
  }

  \item \code{"all"} this gives test results for all theoretically valid
  combinations of bootstrap resampling schemes.

  \item \code{"all and also invalid"} this gives test results for all possible
  combinations of bootstrap resampling schemes and test statistics, including
  invalid ones.
}
A warning is raised if the given combination of \code{type_boot_user} and
\code{type_stat_user} is theoretically invalid.}
}
\value{
A class object with components \itemize{
   \item \code{pvals_df}: a dataframe of p-values and bootstrapped test statistics:

   These are the p-values for the 8 combinations of bootstrap resampling schemes
   (nonparametric and independent), test statistics (centered and equivalent),
   and Kolmogorov-Smirnov or L2-type of true test statistic.

   It also contains the vectors of bootstrap test statistics
   for each of the combinations.

   \item \code{true_stats} a named vector of size 2 containing the true test
   statistics for the L2 and KS distances.

   \item \code{nBootstrap} Number of bootstrap repetitions.

   \item \code{nameMethod} string for the name of the method used.
}
}
\description{
Perform a hypothesis test of statistical independence by means of bootstrapping.
The null hypothesis is that of independence between the two random variables,
versus the alternative of dependence between them.
This procedure gives a total of 8 combinations of bootstrap resampling schemes
(nonparametric and independent), test statistics (centered and equivalent),
and Kolmogorov-Smirnov or L2-type of true test statistic. This function
gives the corresponding p-values, the true test statistic and the
bootstrap-version test statistics. The default (and valid) method implemented
in this function is the null bootstrap, together with the equivalent test
statistic and Kolmogorov-Smirnov test statistic.
Via the \code{bootstrapOptions} argument, the user can specify other
bootstrap resampling schemes and test statistics.
}
\examples{
n <- 100

# Under H1
X1 <- rnorm(n)
X2 <- X1 + rnorm(n)
result <- perform_independence_test(
   X1, X2, nBootstrap = 100,
   bootstrapOptions = list(type_boot = "indep",
                           type_stat = "eq",
                           norm_type = "KS") )
print(result)
plot(result)

# Under H0
X1 <- rnorm(n)
X2 <- rnorm(n)
result <- perform_independence_test(X1, X2, nBootstrap = 100)
print(result)
plot(result)

}
\seealso{
\code{\link{perform_GoF_test},\link{perform_regression_test}}
}
