% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrapTest.R
\name{plot.bootstrapTest}
\alias{plot.bootstrapTest}
\title{Plot the bootstrap test statistics distribution}
\usage{
\method{plot}{bootstrapTest}(
  x,
  xlim = NULL,
  breaks = NULL,
  legend.x = NULL,
  legend.y = NULL,
  ask = interactive(),
  ...
)
}
\arguments{
\item{x}{an object of class \code{bootstrapTest_independence} or \code{bootstrapTest}}

\item{xlim}{limits for the x-axis of the histogram}

\item{breaks}{breaks for the histogram}

\item{legend.x}{position of the legend on the x-axis}

\item{legend.y}{position of the legend on the y-axis}

\item{ask}{if \code{TRUE}, the user is asked to press Return to see the next
plot. Used only if \code{x} is an object of class \code{bootstrapTest_regression}.}

\item{...}{additional arguments passed to the \code{hist} function}
}
\description{
In the regression test case, the estimated regression line is plotted as well.
}
