% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_GoF_test.R
\name{perform_GoF_test}
\alias{perform_GoF_test}
\title{Perform a GoF test}
\usage{
perform_GoF_test(X_data, parametric_fam = "normal", nBootstrap)
}
\arguments{
\item{X_data, }{numerical vector of the same size. the GoF tests whether this
sample comes from `parametric_fam`, a specified parametric distribution.}

\item{parametric_fam}{name of the parametric family. For the moment, only
`normal` is supported.}

\item{nBootstrap}{number of bootstrap repetitions.}
}
\value{
A list with components \itemize{
   \item \code{pvals_df} df of p-values and bootstrapped test statistics:

   These are the p-values for the combinations of bootstrap resampling schemes
   (nonparametric and null), test statistics (centered and equivalent),

   It also contains the vectors of bootstrap test statistics
   for each of the combinations.

   \item \code{true_stat} a named vector of size 1 containing the true test
   statistic.
}
}
\description{
This function performs a goodness-of-fit test for a specific univariate
parametric family.
}
\examples{
n = 100

# Under H1
X_data = rgamma(n,2,3)
result = perform_GoF_test(X_data, nBootstrap = 30)
result$pvals_df
#
# Under H0
X_data = rnorm(n)
result = perform_GoF_test(X_data, nBootstrap = 30)
result$pvals_df

}
\seealso{
\code{\link{perform_regression_test},\link{perform_independence_test}}
}
