% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_GoF_test.R
\name{perform_GoF_test}
\alias{perform_GoF_test}
\title{Perform a GoF test}
\usage{
perform_GoF_test(
  X_data,
  parametric_fam = "normal",
  nBootstrap = 100,
  bootstrapOptions = NULL
)
}
\arguments{
\item{X_data, }{numerical input vector. Perform a GoF test whether or not this
sample comes from `parametric_fam`, a specified parametric distribution.}

\item{parametric_fam}{name of the parametric family. For the moment, only
\code{"normal"} is supported.}

\item{nBootstrap}{numeric value of the number of bootstrap resamples. Defaults
to 100.}

\item{bootstrapOptions}{This can be one of \itemize{
  \item \code{NULL}

  \item a list with at most 3 elements names \itemize{
        \item \code{type_boot} defaults to the \code{"null"} bootstrap
        resampling scheme to be used. \code{type_boot_user} can be either
        \code{"null"} for the null/parametric bootstrap, or \code{"NP"} for the
        non-parametric bootstrap.

        \item \code{type_stat} defaults to \code{"eq"} for the type of test
        statistic to be used. This can be either \code{"eq"} for the
        equivalent test statistic, or \code{"cent"} for the centered
        test statistic.

        \item \code{param_bs} defaults to \code{"canonical"} for the bootstrap
        parameter estimator to be used. \code{param_bs} can be either
        \code{"MD"} for the Minimum Distance estimator, \code{"MD-cent"} for
        the centered Minimum Distance estimator, or \code{"canonical"}
        for the canonical estimator (empirical mean and variance).
  }
  #'   \item \code{"all"} this gives test results for all theoretically valid
  combinations of bootstrap resampling schemes.

  \item \code{"all and also invalid"} this gives test results for all possible
  combinations of bootstrap resampling schemes and test statistics, including
  invalid ones.
}
A warning is raised if the given combination of \code{type_boot},
\code{type_stat}, and \code{param_bs} is theoretically invalid.}
}
\value{
A class object with components \itemize{
   \item \code{pvals_df} df of p-values and bootstrapped test statistics:

   These are the p-values for the combinations of bootstrap resampling schemes,
   test statistics (centered and equivalent), and different parameter estimators.

   It also contains the vectors of bootstrap test statistics
   for each of the combinations.

   \item \code{true_stat} a named vector of size 2 containing the true test
   statistics. The first entry is the Kolmogorov-Smirnov test statistic for
   the Minimum Distance estimator, and the second entry is the Kolmogorov-Smirnov
   test statistic for the canonical parameter estimator.

   \item \code{nBootstrap} number of bootstrap repetitions.

   \item \code{nameMethod} string for the name of the method used.

}
}
\description{
This function performs a bootstrap goodness-of-fit test for a specific univariate
parametric family. It implements a null bootstrap and a non-parametric bootstrap.
The test statistic is the Kolmogorov-Smirnov test statistic.
For the null/parametric bootstrap, the minimum distance estimator is used to
estimate the parameters, or a canonical estimator (the sample mean and variance).
For now, only a test of normality is implemented. This function gives the
corresponding p-values, the true test statistic and the bootstrap-version
test statistics.
}
\examples{
n = 100
# Under H1
X_data = rgamma(n,2,3)
result = perform_GoF_test(X_data,
                          nBootstrap = 30,
                          bootstrapOptions = list(type_boot = "null",
                                                  type_stat = "eq",
                                                  param_bs = "canonical")
                         )
print(result)
plot(result)

# Under H0
X_data = rnorm(n)
result = perform_GoF_test(X_data, nBootstrap = 30)
print(result)
plot(result)

}
\seealso{
\code{\link{perform_regression_test},\link{perform_independence_test}}
}
