% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_independence_test.R
\name{perform_independence_test}
\alias{perform_independence_test}
\title{Perform a test of  independence}
\usage{
perform_independence_test(
  X1,
  X2,
  my_grid = NULL,
  nBootstrap = 100,
  bootstrapOptions = NULL
)
}
\arguments{
\item{X1, X2}{numerical vectors of the same size. The independence test tests
whether \code{X1} is independent from \code{X2}.}

\item{my_grid}{the grid on which the empirical CDFs are estimated.}

\item{nBootstrap}{number of bootstrap repetitions.}

\item{bootstrapOptions}{This can be one of \itemize{
  \item \code{NULL}

  \item a list with at most 3 elements names \itemize{
        \item \code{type_boot_user} type of bootstrap to resample the data,
        either 'NP' or 'cent'.

        \item \code{type_stat_user} type of test statistic to use,
        either 'cent' for centered or 'eq' for equivalent.

        \item \code{norm_type_user} type of norm to use for test statistic,
        either 'L2' or 'KS'.
  }

  \item \code{"all"}:

  \item \code{"all and also wrong"}:
}
A warning is raised if the given combination of \code{type_boot_user} and
\code{type_stat_user} is theoretically invalid.}
}
\value{
A class object with components \itemize{
   \item \code{pvals_df}: df of p-values and bootstrapped test statistics:

   These are the p-values for the 8 combinations of bootstrap resampling schemes
   (nonparametric and independent), test statistics (centered and equivalent),
   and Kolmogorov-Smirnov or L2-type of true test statistic. The column 'bootstrapped_tests'
   contains vectors of bootstrap test statistics.

   \item \code{true_stats} a named vector of size 2 containing the true test
   statistics for the L2 and KS distances.

   \item \code{nBootstrap} Number of bootstrap repetitions.

   \item \code{highlighted_pval} a dataframe with the default independence
   results.
}
}
\description{
Perform a hypothesis test of statistical independence by means of bootstrapping.
The null hypothesis is that of independence between the two random variables.
This procedure gives a total of 8 combinations of bootstrap resampling schemes
(nonparametric and independent), test statistics (centered and equivalent),
and Kolmogorov-Smirnov or L2-type of true test statistic. This function
gives the corresponding p-values, the true test statistic and the
bootstrap-version test statistics.
}
\examples{
n = 100

# Under H1
X1 = rnorm(n)
X2 = X1 + rnorm(n)
result = perform_independence_test(X1, X2, nBootstrap = 30)

result = perform_independence_test(
   X1, X2, nBootstrap = 30,
   bootstrapOptions = list(type_boot_user = "indep",
                           type_stat_user = "eq",
                           norm_type_user = "KS") )

print(result)
plot(result)

#
# Under H0
X1 = rnorm(n)
X2 = rnorm(n)
result = perform_independence_test(X1, X2, nBootstrap = 30)
print(result)
plot(result)

}
