% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_independence_test.R
\name{perform_independence_test}
\alias{perform_independence_test}
\title{Perform a test of  independence}
\usage{
perform_independence_test(X1, X2, my_grid = NULL, nBootstrap)
}
\arguments{
\item{X1, X2}{numerical vectors of the same size. The independence test tests
whether \code{X1} is independent from \code{X2}.}

\item{my_grid}{the grid on which the empirical CDFs are estimated.}

\item{nBootstrap}{number of bootstrap repetitions.}
}
\value{
A list with components \itemize{
   \item \code{pvals_df}: df of p-values and bootstrapped test statistics:

   These are the p-values for the 8 combinations of bootstrap resampling schemes
   (nonparametric and independent), test statistics (centered and equivalent),
   and Kolmogorov-Smirnov or L2-type of true test statistic. The column 'bootstrapped_tests'
   contains vectors of bootstrap test statistics.

   \item \code{true_stats} a named vector of size 2 containing the true test
   statistics for the L2 and KS distances.
}
}
\description{
Perform a hypothesis test of statistical independence by means of bootstrapping.
This procedure gives a total of 8 combinations of bootstrap resampling schemes
(nonparametric and independent), test statistics (centered and equivalent),
and Kolmogorov-Smirnov or L2-type of true test statistic. This function
gives the corresponding p-values, the true test statistic and the
bootstrap-version test statistics.
}
\examples{
n = 100

# Under H1
X1 = rnorm(n)
X2 = X1 + rnorm(n)
result = perform_independence_test(X1, X2, nBootstrap = 100)
result
#
# Under H0
X1 = rnorm(n)
X2 = rnorm(n)
result = perform_independence_test(X1, X2, nBootstrap = 100)
result

}
