% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_independence_test.R
\name{perform_independence_test}
\alias{perform_independence_test}
\title{Perform a test of independence}
\usage{
perform_independence_test(X1, X2, nBootstrap)
}
\arguments{
\item{X1, X2}{numerical vectors of the same size. The independence test tests
whether X1 is independent from X2.}

\item{nBootstrap}{number of bootstrap repetitions.}

\item{my_grid}{the grid on which the empirical CDFs are estimated.
TODO: implement a different grid for X1 and X2.
TODO: can the grid be chosen automatically? For example quantiles of X1, X2?}
}
\value{
A list with components \itemize{
   \item
   # pvals_df: df of p-values and bootstrapped test statistics:

   pval_NP_eq_L2, pval_NP_cent_L2,
   pval_indep_cent_L2, pval_indep_eq_L2,
   pval_NP_eq_KS, pval_NP_cent_KS,
   pval_indep_cent_KS, pval_indep_eq_KS.
   These are the p-values for the 8 combinations of bootstrap resampling schemes
   (nonparametric and independent), test statistics (centered and equivalent),
   and Kolmogorov-Smirnov or L2-type of true test statistic,

   stat_st_NP_cent_L2, stat_st_NP_eq_L2,
   stat_st_indep_cent_L2, stat_st_indep_eq_L2,
   stat_st_NP_cent_KS, stat_st_NP_eq_KS,
   stat_st_indep_cent_KS, stat_st_indep_eq_KS.
   These are the vectors of bootstrap test statistics for each of the combinations.

   \item true_stats, a named vector of size 2 containing the true test
   statistics for the L2 and KS distances.
}
}
\description{
Perform a test of independence
}
\examples{
n = 100

# Under H1
X1 = rnorm(n)
X2 = X1 + rnorm(n)
result = independence_test(X1, X2, nBootstrap = 100)
result$pvals_df
#
# Under H0
X1 = rnorm(n)
X2 = rnorm(n)
result = independence_test(X1, X2, nBootstrap = 100)
result$pvals_df

}
