% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_regression_test.R
\name{perform_regression_test}
\alias{perform_regression_test}
\title{Performs a test on the coefficient of a univariate linear regression}
\usage{
perform_regression_test(X, Y, nBootstrap)
}
\arguments{
\item{X}{numeric input vector}

\item{Y}{numeric input vector}

\item{nBootstrap}{numeric value of the amount of bootstrap resamples}
}
\value{
a named list with `pvals_df` the df of p-values, with the vectors of
bootstrapped test statistics T_n^* included and `true_stat` the true test statistic.
}
\description{
This function performs the bootstrap regression test for given data X,Y. It
uses indep_bs, empirical_bs, res_bs, fixed_design_bs_Hnull, fixed_design_bs,
hybrid_null_bs as bootstrap resampling schemes to perform the bootstrap and
stores all the corresponding p-values and bootstrapped test statistics in a
dataframe to keep track of the outputs.
}
\seealso{
\code{\link{perform_GoF_test}()}
}
