% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrapTest.R, R/print_bootstrapTest.R
\name{plot.bootstrapTest}
\alias{plot.bootstrapTest}
\alias{print.bootstrapTest}
\title{Plot and print the bootstrap test statistics distribution}
\usage{
\method{plot}{bootstrapTest}(
  x,
  xlim = NULL,
  breaks = NULL,
  legend.x = NULL,
  legend.y = NULL,
  ask = interactive(),
  ...
)

\method{print}{bootstrapTest}(x, ...)
}
\arguments{
\item{x}{an object of class \code{bootstrapTest_independence} or \code{bootstrapTest}}

\item{xlim}{limits for the x-axis of the histogram}

\item{breaks}{breaks for the histogram}

\item{legend.x}{position of the legend on the x-axis}

\item{legend.y}{position of the legend on the y-axis}

\item{ask}{if \code{TRUE}, the user is asked to press Return to see the next
plot. Used only if \code{x} is an object of class \code{bootstrapTest_regression}.}

\item{...}{additional arguments passed to the \code{hist} function
(in the case of the \code{plot} method) or ignored (in the case of the
\code{print} method).}
}
\value{
These functions have no return value and are called solely for their
side effects.
}
\description{
The \code{plot} and \code{print} methods work for objects of class \code{bootstrapTest},
or for the specific classes \code{bootstrapTest_independence}, \code{bootstrapTest_GoF},
and \code{bootstrapTest_regression}. The \code{print} method prints the
summary of the bootstrap test results. The \code{plot} method plots the distribution of
bootstrapped test statistics as a histogram, with the true test statistic
and the 95% quantile of the bootstrapped test statistics highlighted.
In the regression test case, the estimated regression line is plotted as well.
}
\seealso{
The functions that generate such object \code{x}:
\code{\link{perform_independence_test}, \link{perform_GoF_test},
\link{perform_regression_test}}.
}
