% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_bootstrapTest.R
\name{print.bootstrapTest}
\alias{print.bootstrapTest}
\title{Print the bootstrap test results}
\usage{
\method{print}{bootstrapTest}(x, ...)
}
\arguments{
\item{x}{an object of class \code{bootstrapTest_independence},
\code{bootstrapTest_GoF}, \code{bootstrapTest_regression} or \code{bootstrapTest}}

\item{...}{additional arguments passed to the \code{print} function}
}
\description{
Print the bootstrap test results
}
